Make the max_episode_steps be an integer (baselines algos expects an int)

+++ gym_wind_turbine/__init__.py
@@ -4,7 +4,7 @@ from gym.envs.registration import register
 register(
     id='WindTurbine-v0',
     entry_point='gym_wind_turbine.envs:WindTurbine',
-    max_episode_steps=60.0/(1.0/20),  # 60s -> 2400 steps
+    max_episode_steps=int(60.0/(1.0/20)),  # 60s -> 1200 steps
     kwargs={
         'env_settings': {
             'timestep': 1.0/20.0,
@@ -20,7 +20,7 @@ register(
 register(
     id='WindTurbineStepwise-v0',
     entry_point='gym_wind_turbine.envs:WindTurbine',
-    max_episode_steps=120.0/(1.0/20),  # 120s -> 4800 steps
+    max_episode_steps=int(120.0/(1.0/20)),  # 120s -> 2400 steps
     kwargs={
         'env_settings': {
             'timestep': 1.0/20.0,
